Added Arduino "ready to use" configuration files
This configuration has not been tested yet. I created initially because I needed for a project, but the project has changed before I had a chance to use it. I decided to put it here because so, in case I was needed again in the future, I'll know where to find it. Therefore, if anyone would like to test it with Arduino would be great.